Replace calls to Get_Indx with Spc_GetIndx
We now use the fast-species lookup routine Spc_GetIndx from the species
database to obtain the species ID. The routine Get_Indx was set up a long
time ago and because of the species database it is no longer needed.

Also removed a parallelized DO loop from flex_chemdr.F. It is not necessary
to parallelize a DO loop over the number of species because there are fewer
than 200 species.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>