Restore Spec_Name, Spec_Id, Trac_Name, and Trac_ID to gigc_state_chm_mod.F90
For an unknown reason, removing the Spec_Name, Spec_Id, Trac_Name, and
Trac_ID fields from the State_Chm object cause the model to crash with
a segmentation fault in restart_mod.F even though they do not appear to be
used anywhere in the code. Compiling with any debug flags turned on does
not return an error, so debugging this problem is difficult at this time.
The fields are restored for now to avoid errors, but we can attempt to
remove them again once the code is in a more stable state.

Also restored the initialization of State_Chm%Trac_Units  = ''. This was
likely removed by accident in an earlier commit.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>