Set species ID using Spc_GetIndx as a temporary workaround
The species IDs used the ordering set by SMVGEAR, but the code now assumes
the order set by the species database. To ensure we're getting the proper
species information, we now set the species IDs (IDxxxx) using Spc_GetIndx
in routine SETTRACE (found in GeosCore/tracerid_mod.F). This is a temporary
fix because the species ID are slated for removal in the near future.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>