Add fix to not overwrite family tracer concentrations after chemistry
The STT values for the family tracers are set in the FAMILIES_KLUDGE
routine before we loop over species and copy the updated State_Chm%Species
values into STT. The STTTOCSPEC .NE. 0 check for skipping family tracers
doesn't work anymore because all of the family tracers are defined in the
species database and  thus have a STTTOCSPEC value. Instead, to avoid
overwriting STT for the family tracers, we need to explicitly check that
Input_Opt%Tracer_Name isn't one of the family tracer names.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>