Add bug fix for error messages in geosfp_read_mod.F90 and merra2_read_mod.F90
When the GEOS-FP and MERRA-2 code was updated to open files once per day,
several calls to Error_Stop that utilized the string caller were moved
ahead of the lines that defines caller. To resolve this, the line defining
caller, or the name of the current routine, is moved to the top of each
routine.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>