Replace loop over NSPEC with check for Is_KPP in flexchem_setup_mod.F90
In flexchem_setup_mod.F90, there was a loop over KPP species (NSPEC) added
for the sole purpose of checking if the GEOS-Chem species a KPP species
before converting background initial concentrations from mol/mol to
molec/cm3. For a cleaner implementation, we can replace the loop over
NSPEC with an IF statement to check if it's a KPP species using
ThisSpc%Is_Kpp. Difference tests showed that this update produces identical
results to the prior code with the NSPEC loop.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>