Add temporary fix to set dead species CO2 to zero in flex_chemdr.F
We need to set CO2 to zero to match the old SMVGEAR species code. Because
CO2 is defined as D (dead) in globchem.dat, the old SMVGEAR code would
zero concentrations for that species automatically. Now, CO2 is a variable
species in KPP and is set to the background value hardcoded in restart_mod.F.
To avoid differences for now, we will set CO2 to zero.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>