Comment out code in diag_pl_mod.F - this diagnostic is broken
To avoid issues while removing dependencies on SMVGEAR variables, the
code in diag_pl_mod.F has been completely commented out. The prod/loss
diagnostic is broken right now because it was highly dependent on SMVGEAR
and was never set up to work with KPP. Mike Long has added the capability
to save out prod/loss information from KPP, but this still needs to be
connected to GEOS-Chem.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>