Further optimizations to Rachel's codeathon updates:
benchmark_mod.F: Now use the pre-defined variable ind_O3 in the
call to the BPCH2 routine as the tracer ID number.

diag3.F: Because the IND_() routine returns -1 instead of 0 if it cannot
find a species, change the id_SOAG /= 0 to id_SOAG > 0 and ditto for
id_SOAM.

main.F: Now pre-save the value of IND_('CLOCK') in the id_CLOCK
variable.  Also now use ind_CLOCK > 0 as the criterion to call
SET_CLOCK_TRAC, because IND_() returns -1 if not found.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>