Add Is_Photolysis, Is_ActiveChem, and Is_FixedChem fields to species database
The Is_ActiveChem and Is_FixedChem are set according to the KPP indices.
If KppFixId is > 0 then it's a fixed chemical species. If KppVarId is > 0
and KppFixId is <= 0 then it's an active chemical species. These flags
should help identify inactive that need special handling in the chemistry
code (e.g. flex_chemdr.F, ucx_mod.F).

The Is_Photolysis field is also added. In routine Init_Species_Database,
the Is_Photolysis flag is now set to T for several of the tropchem species.
Because many of the FAST-JX photolysis species are not already defined in
the species database (i.e. they are not tracers, but KPP species), several
new species have been added with CASE statements to the species list so
that Is_Photolysis can be set to T for these species.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>