Move some error checks after Species Database initialization
Some of the error checks (in READ_AEROSOL_MENU and READ_EMISSIONS_MENU)
now use the IND_() function in gigc_state_chm_mod.F90 to determine if
a species is defined before using a specific option.  But many of these
error checks used to rely on tracerid_mod.F, which was called as soon
as the list of tracers was read in.

Therefore, we have now moved these error checks to routine
DO_ERROR_CHECKS, which is now called from GIGC_INIT_EXTRA.  This will
make sure that the species database is initialized before we attempt
to use the IND_() function to look up species indices.

Also note: IND_() returns -1 for "species not found", but the existing
error checks used 0 for "species not found".  Therefore, we now force
-1's to 0's in certain error checks so that the prior program logic
is preserved.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>