Bug fix in diag_oh_mod.F: Only define id_OH for fullchem sims
CH4 simulations do not have OH as a defined species.  Therefore
the error check for id_OH < 0 from the prior commit caused CH4
simulations to die with an error.

We have now moved the definition and error check of id_OH to within
the IF ITS_A_FULLCHEM_SIM block.  This corrects the error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>