Remove references to APM; it's no longer compatible with FlexChem
The GeosApm directory was based on v9-01-02 or v9-01-03.  Since then,
we have made many sweeping changes (i.e. derived type objects, HEMCO,
FlexChem, removal of SMVGEAR) to GEOS-Chem to facilitate running in HPC
environments.  But APM was not kept up to date.  With the removal of
tracerid_mod.F, the APM code is now irretrievably broken.

The best thing to do is to remove it totally from GEOS-Chem, so that
it can be cleanly re-implemented in a future version.

Also: Remove the file GeosCore/tracerid_mod.F.  We should have done
this a couple of commits ago.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>