Set Is_Photolysis for UCX species in the species database
The previous commit added the Is_Photolysis field and set it for several
of the tropchem species. Here, Is_Photolysis is also set for the UCX
species. For species found in both tropchem and the UCX that undergo
photolysis in UCX but not in tropchem (i.e. NO, SO4, CH2Br2, CH3Br), the
line for Is_Photolysis had been blocked off by #if def(UCX). This line is
not blocked off for species that only occur in the UCX mechanism (e.g.
Cl species, N2O, OCS).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>