Add an #if defined( ESMF ) block to CLEANUP_GIGC_STATE_MET
In the previous commit, several DEALLOCATE statements were changed to
NULLIFY statements to prevent an error in GCHP.  We have now wrapped
these statements in an "#if defined( ESMF )" block, which is only
activated for GCHP simulations.  This will preserve the prior behavior
for GEOS-Chem "Classic" simulations.

Usually, if you allocate memory to a pointer variable, you also need
to deallocate it, in order to avoid a memory leak.  If you point to
a target varaiable, then the pointer can be nullified.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>