Remove last vestiges of SMVGEAR: comode_loop_mod.F, readchem.F, and reader.F
The previous commits have paved the way for the removal of comode_loop_mod.F.
Here, the final steps in that removal are done. Calls to initialize and
cleanup comode_loop_mod.F are removed. The comode_loop_mod.F file is deleted
and removed from Headers/Makefile.

The removal of comode_loop_mod.F, also allows for the removal of readchem.F
and reader.F and the sparse-matrix routines that they utilize (jsparse.F
and ksparse.F). These routines are no longer needed because we now utilize
KPP entirely for the chemistry solver and we obtain species information from
the species database. This also means that the globchem.dat and mglob.dat
files are no longer utilized.

This commit should mark the complete removal of SMVGEAR code from GEOS-Chem.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>