Remove references to STT and State_Chm%Tracers (Bob's updates #1)
We have started removing references from State_Chm%TRACERS (and the
local STT pointers that are used in many routines).  We have started
focusing on the specialty simulations,

We eventually want to use State_Chm%Species everywhere.  During the
development, we are forcing State_Chm%Species = State_Chm%Tracers
in individual subroutines (and vice versa before exiting) so that
we do not break any functionality.  This can be removed once the
State_Chm%TRACERS is removed from all routines everywhere.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>