Remove references to STT and State_Chm%Tracers (Bob's updates #2)
Continued replacing references to State_Chm%Tracers with
State_Chm%Species.  Also replace STT with Spc for clarity.

In a few routines, we also renamed the Species Database object
ThisSpc to SpcInfo, in order to reduce confusion with Spc.

Also, for now, commented out the call to CHECK_STT in main.F
while testing is ongoing.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>