Remove references to STT and State_Chm%Tracers (Bob's updates #3)
Continued replacing references to State_Chm%Tracers with
State_Chm%Species.  Also replace STT with Spc for Clarity.

This commit includes most of the remaining specialty simulations
(except for Hg, to which I need to return later), and the diagnostics.

NOTE: I did not thoroughly test the timeseries diagnostics, because
they are in bpch format (which will be retired shortly).

ALSO: We have commented out the calls to CHECK_STT.  These are obsolete
and could possibly be retired (compiling with FPE=y also performs
checks for floating-point errors.)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>