Fixed bug in sulfate_mod.F that was causing floating-point errors
In routine CHEMSULFATE, we had to move a temporary copy of
State_Chm%Tracers to State_Chm%Species so that it was outside of
an IF ( ITS_AN_AEROSOL_SIM ) block.  This was causing errors in
a benchmark unit test.  Now fixed.

This modification passed an aerosol diff test (11 Jul 2016).

NOTE: Need to check to see why we do not get identical results
in a tropchem diff test.  We might be lacking blocks that copy
species <-> tracers in some key places yet.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>