Add Mike's updates for obtaining prod/loss rates from KPP
Updated flex_chemdr.F to include a call to routine Flux, which is found in
gckpp_Function.F90. This will return arrays PROD and LOSS, which can be
used by specifying the KPP index of the desired prod/loss species (e.g.
PROD_CO2 = PROD(IND_CO2)).

The deltaCheck array appears to be for debugging purposes only and can
be removed later. Also, a minor update was made to the KPP diagnostics to
change dimensions from 3D to 2D.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>