Remove references to STT and State_Chm%Tracers (Bob's updates #9)
Replaced references to State_Chm%TRACERS with State_Chm%SPECIES
in GeosCore/transport_mod.F and GeosCore/vdiff_mod.F90.

IMPORTANT NOTE: For now we still assume that the advected species
are listed first.  This allows us to just minimally change the
pointer assignments, e.g. from:

   p_STT => State_Chm%TRACERS( :, :, LLPAR:1:-1, : )

to

   p_Spc => State_Chm%Species( :, :, LLPAR:1:-1, 1:nAdvect )

This also lets us avoid having to rewrite the internal routines
of the transport and non-local PBL mixing modules.

BUT!  We may have to revisit this at some point in the future,
if we ever decide to reorganize the species listing such that the
advected species no longer all come first.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>