Replace tracer pointer STT with species pointer SPC locally in ucx_mod.F
NOTES:
(1) All routines locally save the initial species values prior to
setting species equal to tracers. The original species values are restored
at the end of the routine.

(2) Routine settle_strat_aer continues to use tracers instead of species
in this commit, although the code to transition to species is there but
commented out. Pointing to species instead of tracers results in
failure of the UCX difftest, with differences in HEMCO and GC restarts and
diagnostics.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>