Move tracer removal modification to just prior to settle_strat_aer call.
When UCX is enabled, settle_strat_aer is called within sulfate_mod.
Currently tracer removal is not working properly in settle_strat_aer.
As a result, the pointer to tracers remains in that function in order
to pass difference tests. In order for it to work properly, however,
the recently added tracer removal modification in sulfate_mod must
end prior to the settle_strat_aer call. Moving it to prior to the call
restores difference tests to passing.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>