Make sure HCFC141b and HCFC142b retain a lowercase b in the species name
The HCFC141b and HCFC142b species need to be hardcoded to use
NAME = 'HCFC141b' and NAME = 'HCFC142b' in routine Init_Species_Database.
Using Name = NameAllCaps causes problems when trying to match the
species name in other locations, such as in restart_mod.F when trying
to match the SPC_ field from the restart file with the species name.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>