Add fixes for CFC and HCFC species in SET_INITIAL_MIXRATIOS
In the calls to APPLY_2DTRAC, we no longer need to pass TRAC_ADD=.TRUE.
for CFC114, CFC115, HCFC141b, and HCFC142b. These are no longer members
of a family, but are their own species so they do not need to be summed.

Also fixed the species ID passed in the IF statement for CFC115, HCFC141b,
and HCFC142b in SET_INITIAL_MIXRATIOS. These were using CFC113 and HCFC123,
but this was likely a copy/paste error.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>