Bug fix: revert to using State_Chm%TRACERS in ExtState_SetFields
We discovered that (for now at least) we need to pass the
State_Chm%Tracers fields for O3, NO2, NO, HNO3 and POPG to the
ExtDat_Set routine.  Otherwise, State_Chm%Species is in the wrong
units when PARANOX is called and will return erroneous PARANOX
emission fluxes.

This will be resolved at the very end when we remove the
State_Chm%Tracers fields everywhere from GEOS-Chem.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>