Routine GET_O3 now does the proper unit conversion
Routine GET_O3 (in GeosCore/sulfate_mod.F) previously assumed that
State_Chm%Species was in molec/cm3 and thus applied a unit
conversion to v/v.  In older versions of GEOS-Chem, O3 was not
an advected species (Ox was), so we had to look up O3 from the
chemical species array State_Chm%Species (formerly known as
CSPEC_FULL.

The removal of State_Chm%Species means that by the time the program
flow gets to the GET_O3 routine, the units are already in v/v.
Thus, the unit conversion is no longer necessary.