Remove use of State_Chm%Tracers from diag20_mod.F
In preparation for the removal of State_Chm%Tracers, the uses of pointer
STT (which points to State_Chm%Tracers) have been removed from routine
DIAG20. STT was used to obtain the O3 mass in [kg], which is converted to
[molec] for the LOx calculation. The routine now obtains O3 concentrations
from State_Chm%Species [molec/cm3] and converts to [molec] by multiplying
by BOXVL [cm3]. This update results in identical ND20 output for LOx.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>