Rename "tagged Ox" to "tagged O3"
The name "tagged Ox" was historical. The Ox tracer in GEOS-Chem was
replaced with O3 in v9-02. Previously, the nomenclature "tagged Ox"
was interchangeable with "tagged O3", but now "tagged O3" is used
throughout the code for clarity. As part of this update, tagged_ox_mod.F
has been renamed to tagged_o3_mod.F.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>