Make sure species go into flex_chemdr in consistent units
We have had to convert State_Chm%Species to kg/kg dry first, before
copying in the non-advected species.  We have verified that the
species go into FLEXCHEM in the same units as in the Reference
code.

NOTE: We may still be getting numerical differences in this
commit.  We are still looking into this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>