Fix bug where new family species background values not set above LLCHEM
With this update, background values used when species are not in the
restart file stored in the species database are now assigned to
State_Chm%Species for ALL levels for advected species, and for levels
up to LLCHEM for non-advected species.  For levels above LLCHEM,
non-advected species are set to a small number.

Also in this commit:

    (1) during species initialization, when writing out molec/cm3
        in debug mode, take the min and max of all levels, not just
        1:LLCHEM.

    (2) add temp routine to print out tracer and advected species values
        in kg/kg dry as well as all species values in molec/cm3. This
        is a temporary routine for use in tracer removal. It is stored
        in restart_mod and can only be called when tracers are in
        kg/kg dry and species are in kg/kg dry or molec/cm3.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>