Update tracer removal code in main.F
This version allows use of species through main.F in place of tracers
in units of kg/kg dry. It has not been tested with flex_chemdr call
enabled, and in fact will not work if that call is turned on in
chemistry_mod. There are several remaining sections of main.F that are
still temporary (debug print calls and copying species into tracers).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>