Remove references to tracers in several routines (#2)
The comments in several diagnostic routines have been updated so that
tracer has been replaced with species where appropriate. Also updated some
of the diagnostics code so that we obtain species name from the species
database instead of Input_Opt%TRACER_NAME.

Also in this commit are fixes for minor bugs in tagged_o3_mod.F and
global_ch4_mod.F. These were necessary to get the code to compile.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>