Remove State_Chm%Tracers and references to tracers from planeflight_mod.F
Uses of STT, which points to State_Chm%Tracers, have been replaced with
Spc, which points to State_Chm%Species. Also, ThisSpc has been renamed to
Spc info for consistency with other routines throughout GeosCore.

NOTE: The use of "tracers" and "TRC_" in the planeflight.log file remains
because we didn't want to break this diagnostic for users wanting to reuse
old planeflight.dat files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>