Remove tracer removal code and references to tracers in tomas_mod.F and dust_mod.F
The TOMAS code in tomas_mod.F and dust_mod.F has been updated to remove
all temporary tracer removal code and any references to tracers.

Also in this commit, we fix a bug where State_Chm%Tracers was still used in routine
GET_DUST_ALK when State_Chm%Species should be used. This was likely the
issue that was causing the aciduptake unit tests to fail.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>