Do not add dummy RR* species from KPP to the species database
For the KPP prod/loss diagnostic a dummy variable starting with 'RR' is
added for each reaction in the chemical mechanism. This results in NREACT
additional KPP species. We do not need to track these species in GEOS-Chem,
so we can skip any KPP species beginning with 'RR' when adding KPP species
to the species database.

This update should help improve model performance. Including the dummy
prod/loss species in the species database was causing extremely slow code,
especially when writing out all species to the GEOS-Chem restart file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>