Remove NESTED_SE option for SEAC4RS / Southeast Asia nested grid
The nested SE grid was originally created for the SEAC4RS campaign because
the 0.25x0.3125 nested China grid did not extend far enough south. However,
it appears that when the SEAC4RS campaign in Southeast Asia was cancelled,
all code updates for this grid were abandoned and so it was never fully
implemented. Here we remove code for the nested SE grid to avoid user
confusion.

In v11-01, we introduce the expanded 0.5x0.625 nested Asia (AS) grid for
MERRA-2, so users will still have the option to use a nested grid that
includes Southeast Asia.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>