Don't call EMISSCARBON unless it's a fullchem or aerosol sim
In routine EMISSIONS_RUN (in GeosCore/emissions_mod.F90), we needed
to wrap the call to EMISSSCARBON in an IF block so that it will
only execute if it is a fullchem or aerosol simulation.

This error was detected with GNU Fortran.  It would halt GEOS-Chem
with an invalid memory error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>