Add bug fix for the CO biomass burning emissions diagnostic
In GeosCore/hcoi_gc_diagn_mod.F, the category passed to
Diagn_Create should be -1 for GFED/FINN emissions and
CATEGORY_BIOMASS for QFED emissions. The category is set
accordingly (via Cat) at the top of routine Diagn_Biomass.
To use the appropriate category value, we need to pass Cat
to Diagn_Create.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>