Prevent array out-of-bounds error in TOMAS simulations
In aero_drydep.F, the code for defining species IDs was blocked off by
an IF statement for FIRST. At the end of that IF block, FIRST was set to
false. This caused issues because there was already an existing IF
statement utilizing FIRST a few lines down and that IF block was no longer
getting executed. As a result, array DRYD was not defined and an array
out-of-bounds error occurred for AD44, which uses DRYD(BIN). To fix this
issue, the two IF blocks for FIRST have been combined.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>