Bug fix: Avoid floating-point error in POPS simulation
In routine CHEM_POPGP (in GeosCore/pops_mod.F), the C_NO3 value
is used in an expression, but is never given a value.  Probably
what happened was that it was assigned a value but then that line of
code was taken out, leaving the orphaned C_NO3 variable.

The GNU Fortran compiler (using FPE=yes) will cause GC to die with
a floating point exception because of this.  The simple fix was just
to set C_NO3 = 0.0_fp.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>