Bug fix: make sure MW_HNO3 is defined for aerosol-only sims
In routine SEASALT_CHEM (in GeosCore/sulfate_mod.F), the MW_HNO3
variable was undefined for aerosol-only simulations, which do not
use HNO3 as a defined species (i.e. where id_HNO3 < 0).

We now set MW_HNO3 to 63.0 for aerosol-only simulations.  This issue
was detected in an aerosol-only unit test using the GNU Fortran
compiler.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>