Don't call DO_RED_INPLUME until after HEMCO_Config.rc is read.
Jaegun Jung (Ramboll Environ) pointed out this error: If the
LInPlume flag is turned on in HEMCO_Config.rc, then the code
will die with an error. This is because we are trying to get
a HEMCO pointer from DO_RED_INPLUME before the HEMCO_Config.rc
file is read.

The fix is to move the call to DO_RED_INPLUME into the routine
SET_OPTIONS_FROM_HEMCO, which is called after the HEMCO_Config.rc
file is read.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>