Bug fix: allocate the OCCONV array for marine POA simualtions
In GeosCore/seasalt_mod.F, the OCCONV module array is used for
marine POA simulations.  It is referenced in routines CHEM_MOPO
and CHEM_MOPI.  However, OCCONV was never allocated.

We added an IF block in INIT_SEASALT to allocate OCCONV only if
we are doing a marine POA simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>