Now set the compile command to mpifort if HPC=yes (or HPC=y)
We have fixed an issue in the Makefile_header.mk.  When HPC=yes or HPC=y
(any combination of caps/lowercase), we set COMPILE_CMD=mpifort if
the Intel compiler is being used.  This will make sure that the
"MPI magic" is activated when we build GEOS-Chem as well as the gridded
component code in the GCHP folder.

Also added a new variable IS_HPC which is 1 if we are compiling for
HPC or 0 if not.  This is used further down in Makefile_header.mk
in other tests for HPC.

NOTE: Right now, the Intel compiler is the only compiler compatible
with GCHP.  This may change in the future.

Signed-off-by: Robert Yantosca <yantosca@clhead.dal.acenet.ca>