Bug fix: Now use proper MW of OH in tagged_co_mod.F
The unit conversion from kg/m3 OH to molec/cm3 OH in a previous
commit (a9ac9e0) had an error.  The conversion was listed as:

  ( AVO / 17.0_fp ) * 1e-6_fp

but should have been

  ( AVO / 17.0e-3_fp ) * 1e-6_fp

The term in parentheses is the old "XNUMOL" factor, which is
( molec/mole ) / ( kg/mole ) = molec/kg.  So the molecular weight
was listed in g/mole but should have been in kg/mole.  This is
now corrected.  Thanks to Bea Bukosa for the heads up.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>