Minor fix in hco_calc_mod.F90 to avoid segmentation fault
The line "Arr2D = 0.0_hp" in HCO_EvalFld_2D was causing a segmentation
fault in the GEOS-Chem unit test simulations when that routine is called
from HCO_SetPBLm. This appears to be caused by HcoState%Grid%PBLHEIGHT%Val
not being associated. To avoid this error, the line for Arr2D is moved to
after the IF statement checking if PBL_HEIGHT is found. A similar fix is
made for Ar3D in routine HCO_EvalFld_3D.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>