Add M_ARCH flag to set the -march setting for GNU Fortran
Now do not set -march flag for GNU Fortran by default.  We now
have added the M_ARCH Makefile variable that allows the user to
pass a value for -march to the build sequence.

NOTE: On some CPU's (e.g. Intel-based), it is better to not set
the -march flag.  This will avoid some assembler errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>