Integrated gas-phase iodine chemistry
The "Halogens" chemistry mechanism has now been extended with
gas-phase chemistry (and photochemistry) for iodine, in
addition to some updates to chlorine and bromine chemistry. This
does not include new heterogeneous chemistry, such as the new
iodine/bromine sea-salt formulations. It was also necessary to
increase the maximum number of cross-sections in Fast-JX, hence
the change to CMN_FJX_MOD.F.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>