Set ACTA and HCOOH to active in the KPP chemical mechanisms
Katie Travis reported an error in the globchem.dat file. Both ACTA and
HCOOH are set to inactive (I), but they undergo chemistry and should be
set to active (A).

In v11-01g, FlexChem was implemented and the globchem.dat file was removed
completely. To set these species as active, they are now added to the
variable species (DEFVAR) list  and removed from the fixed species (DEFFIX)
list in the globchem.spc file that is used by KPP to build a chemical
mechanism. The Standard, Tropchem, UCX, SOA, and SOA_SVPOA mechanisms have
been rebuilt with this fix.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>