Bug fix: move UCX calls to after spc unit conversion to kg in DO_FLEXCHEM
In FlexChem_mod.F90, routines UCX_NOX and UCX_H2SO4PHOT are called when
species concentrations are in molec/cm3. However, both routines expect
concentrations in kg. To correct this issue, the UCX routine calls are
moved to after the species unit conversion to kg at the end of DO_FLEXCHEM.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>