Update Ox family definition for P/L diagnostic
Daniel Jacob, Mat Evans, Barron Henderson, Jenny Fisher, and Tomas Sherwen
gave input on how to define the Ox family in GEOS-Chem. The final decision
is to use:

  Ox : O3 + NO2 + 2NO3 + PAN + PMN + PPN + HNO4 + 3N2O5 + HNO3 + BrO +
       HOBr + BrNO2 + 2BrNO3 + MPN + ETHLN + ISN1 + ISOPNB + ISOPND +
       MACRN + MVKN + PROPNN + R4N2 + INPN + ISNP + INO2 + ISNOOA +
       ISNOOB + ISNOHOO + MAN2 + PRN1 + PRPN + R4N1 + PMNN + MACRNO2 +
       ClO + HOCl + ClNO2 + 2ClNO3 + 2Cl2O2 + 2OClO

For the Tropchem, SOA, and SOA_SVPOA mechanisms, the species ClO, HOCl,
ClNO2, Cl2O2, and OClO are removed from the Ox family because they are not
defined in those mechanisms.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>