Add fix to PM2.5 code to avoid out-of-bounds error in SOA-SVPOA simulations
For SOA-SVPOA simulations, the species OCPI and OCPO are replaced with
POA1, POA2, OPOA1, and OPOA2. Therefore, when computing Rho_dry for use
in ORG_GROWTH, we can't use the density of OCPI. Instead, we use the
density of POA1 which is now set to the same value as OCPI in the species
database.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>