Nullify unused pointers in state_met_mod.F90
Many pointer fields of state_met_mod.F90 may or may not get allocated,
depending on the met fields being used, or if we are running in GCHP
mode.  Make sure to nullify these pointer fields in order to prevent
the compiler (esp. gfortran) from dying with an uninitialized field
error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>