Convert units from kg/kg to kg for species mass output in benchmark_mod.F
In the early tracer unit conversion work, the units in benchmark_mod.F were
changed to kg/kg. However, the benchmark code expects this output to be in units
kg. To fix this, we now convert from kg/kg to kg at the beginning of
routine STDRUN and convert back to kg/kg at the end of the routine.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>