Add temporary fix to get TOMAS to pass unit tests
In TOMAS drydep routine GeosCore/aero_drydep.F, we have commented
out a parallel loop.  The problem is that we are trying to save
3-D data within a parallel loop to the AD44 array, which does not
have a vertical dimension.  This causes roundoff differences which
are enough to make the geosfp_4x5_TOMAS15 and geosfp_4x5_TOMAS40
have differences in the binary punch files (i.e. unit test result
"yellow").

The simple solution was to comment out the parallel loop where
the archiving to ND44 occurs.  We will engage the TOMAS team
for a more permanent solution.

Also: in v11-02, the bpch diagnotics are slated to be removed,
so it might not be worth the trouble to invest a lot of effort into
fixing this issue.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>