Implement temporary new chemical mechanisms for testing
There are now 4 new chemical mechanisms not present in base
GEOS-Chem. These are:

 - NoHet: Standard chemistry without heterogeneous chemistry
 - Hetv2: Identical to standard chemistry, but the heterogeneous
          chemistry mechanism has been modified so that reactions
          previously specified as A + B = C are now specified as
          two half reactions, with A = C and B = O2.
 - Hetv3: As Hetv2, but with bromine and chlorine chemistry updated
          to follow Johan Schmidt's work.
 - Halogens: As Hetv3, but with Tomas Sherwen's iodine chemistry.

This is still a WIP, but Hetv2 and Standard can be directly compared
to show how reformatting the heterogeneous chemistry results in small
changes in (for example) OH. Hetv3 and Halogens are not yet complete,
or even functional.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>