Implemented BrNO3 and ClNO3 hydrolysis in Schmidt halogen scheme
Hydrolysis of BrNO3 and ClNO3 is now implemented based on the scheme
described in Schmidt et al. This scheme results in slower hydrolysis
of BrNO3 but allows hydrolysis of both throughout the atmosphere,
where previously ClNO3 hydrolysis occurred only in the stratosphere.
The groundwork has also been laid for the next reactions to be
updated, specifically HOBr + HBr and HOBr + HCl.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>