Add fixes for compiling with PRECISION=4
GeosCore/wetscav_mod.F: Make sure the locally-defined varaiables
K0, CR, pKa are declared as REAL(f8). This will be consistent with
the species database.

Headers/species_database_mod.F90: Make sure that we pass a _fp
constant to the DD_Hstar_Old field of the species database.

KPP/*/gckpp_precision.F90: We had to add a kludge for compiling with
PRECISION=4.  When you build the KPP files (outside of GEOS-Chem),
they are declared as REAL*8 using the "dp" parameter.  But in
order to "trick" KPP into thinking we can switch the precision
to REAL*4, we have added an #ifdef that will define dp with a
REAL*4 KIND parameter if we compile with PRECISION=4.  This will
all be alleviated once we get GEOS-Chem to build KPP on-the-fly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>