Updated code to allow reactions with BrSALA and BrSALC
Even when we are not explicitly emitting sea-salt bromine, we
allow for ongoing formation and removal of "excess" bromine on
fine and coarse sea salt. This necessitates inclusion of the
BrSALA/BrSALC species and reactions. The code has been extended
to allow for this, and dummy reactions included. This also
results in small changes in ozone loss/production at the top of
the stratosphere, possibly as a result of matrix reordering to
accomodate the new (zero-rate) reactions in KPP.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>