Implemented pre-chemistry call to calculate pH and alkalinity
The new Br/Cl het-chem relies on cloud pH and alkalinity, but these
quantities are not calculated until after the gas-phase chemistry is
called. To get around this, a new call to sulfate_mod has been
inserted into chemistry_mod.F prior to calling chemistry, so that
these quantities are calculated beforehand. The chemsulfate routine
has also been modified so that it can be called in such a way that
only these properties are calculated, without affecting diagnostics
or tracer concentrations. The cloud pH and alkalinity are stored in
variables in the State_Chem state variable. The effect of cloud pH
and alkalinity on het-chem is not currently calculated, for the
purposes of establishing that simply calculating cloud pH does not
affect other chemistry. These variables will be propagated to the
het-chem routines in the next commit.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>